refactor(workflow-runtime): extract worker tests via #[path]
workflow_runtime_worker.rs is the largest "god module" in the harness
codebase at 3076 lines (3.8x U-16's 800-line ceiling) and the bulk of
that — 790 lines — is two `#[cfg(test)] mod tests {...}` blocks. This
PR moves them out via the standard `#[cfg(test)] #[path = "..."] mod
tests;` pattern (same shape used in #1064 for workflow.rs).

Net effect:
  workflow_runtime_worker.rs:                3076 -> 2286 lines (-790)
  workflow_runtime_worker_tests.rs:           NEW   656 lines
  workflow_runtime_worker_profile_tests.rs:   NEW   134 lines

This is Phase 0 PR-1 of the architecture decoupling roadmap (see
docs/stateless-repo-backlog-poll-spec.md and the audit reports). It is
mechanical: zero logic changes, two file moves only. The 4 production
helper functions between the two test mods (last_agent_summary,
last_error, truncate_summary, pretty_json) are preserved in their
original location.

Subsequent Phase 0 PRs will extract:
- workflow_runtime/store.rs tests (~4500-line tests.rs there)
- harness-workflow/runtime/reducer.rs tests
- workflow_runtime_submission.rs tests

Phase 1 PRs (cohesive splits) become possible once worker.rs is
editable under U-16:
- extract workflow_runtime_prompt.rs (~500 lines)
- extract workflow_runtime_workspace.rs (~200 lines)
- which in turn unblocks Issue #1081 (start_child_workflow wire-format
  fix) since the activity_result_schema function moves into a small
  editable file.

Verification:
- cargo test -p harness-server --lib -- workflow_runtime_worker:
    23 passed (tests mod) + 6 passed (profile_tests mod) = 29/29
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean
- cargo fmt --all -- --check: clean

The two new files use `use super::*;` so they see worker.rs's internals
unchanged. No public API surface change. No test logic change.
